+++ _main.R
@@ -1,3 +1,5 @@
+# generate site
+# use_vignette("")
 
 devtools::document()
 pkgdown::build_site()