+++ NEWS.md
@@ -1,5 +1,8 @@
 # RAWdataR 0.5
 
+* fix `raw.findFiles` by MD5 code
+* Validation of files `raw.isValid()`
+* (Documentation with References)[https://thomasgredig.github.io/RAWdataR/]
 * tools library used for MD5 checks
 * use_news_md()
 * Added a `NEWS.md` file to track changes to the package.